api/requirements.txt
# AnesGuardian Vercel 后端依趖
# Vercel 會自動安裝這些依趖

fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
python-multipart>=0.0.20
httpx>=0.28.0
Pillow>=11.0.0
google-generativeai>=0.8.0
